Add a section for events.
This section provides guidelines for working with the Nakadi
API and defining events as part of a service interface. Much of
the current guidelines apply to events (especially around JSON
content). Nakadi itself is defined in terms of a RESTful API, but
the interaction model between producers and consumers is different
in style to a client/server API approach.