Elides references to Nakadi in event guidelines.
Part of series of edits to abstract out the dependency the guidelines have
on the Nakadi project (guidelines should not have hard dependency on
technology implementations). This reduces the number of Nakadi references
without changing the meaning or intent of the guidelines leaving Nakadi
referenced as an example of a system that supports events. The guidelines
still link to the Open API definition hosted by the Nakadi project - how the
formats evolve and where they live is separate work.

The single guideline change is to call out the need for reviews of event
designs as a separate heading - "Make Events available for review". This
is making what's already in the guidelines clearer.

For #237.