Adds Nakadi event categories and event types to the guidelines.
This extracts the event categories and event types defined in the Nakadi into
the guidelines with the intent of allowing the API guidelines to govern and
evolve these formats as de-facto standards for events. Nakadi itself becomes the
reference implementation of the event formats as well as their registration and
validation. Previously the guidelines referenced the Nakadi project's API
definition, which by having design guidelines depend on a technology project was
"the wrong way around". This change allows the API guidelines to inform
technology, evolve the event model over time, and have other services, libraries
and tools draw from guidelines for their model of events.

Some notes:

- Because the guidelines are now responsible for event data models, a brief
  summary of the main concepts is added to the preliminary section.

- Certain information in the event type is specific to Nakadi (such as
  metrics). This has not been brought into the guidelines, but the event type
  format is declared open for extension.

- The Nakadi Business Category has been renamed to the General
  Category. It turns out "business process" was a bit fuzzy as a concept, and
  also teams use its structure as a general purpose "event". Hence the category
  was renamed to reflect how it's being used. A note on the old name has been
  added. About 90% of Nakadi's registered event types within Zalando use the
  business process category.

- The undefined event category has not been carried forward into the guidelines
  since it's not strictly speaking, an event. There's no place to declare
  metadata for example, and an undefined event is really just an object
  conforming to a schema. The undefined event still exists in Nakadi and there
  are no plans to remove it, because it's useful for some usecases (as a
  datapoint, less than 5% of Nakadi's registered even types within Zalando are
  declared undefined).

- In all cases, the object definitions here are taken as-is from Nakadi with no
  changes in structure or semantics. Practically speaking this is "as you were"
  with respect to the impact guidelines have on day to day work with events and
  the aim also is not require major work in Nakadi (the single exception would
  be to support the category name "general" as well as the previous name
  "business").

The guidelines will evolve event structures in the future as there is a desire to
do so, but for now the goal is to have events as they are defined and used today,
as part of the guidelines.

For #237.